+++ src/Endpoints/ShippingGroups/Find.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace Hitmeister\Component\Api\Endpoints\ShippingGroups;
+
+use Hitmeister\Component\Api\Endpoints\AbstractEndpoint;
+use Hitmeister\Component\Api\Endpoints\Traits\RequestGet;
+
+/**
+ * Class Find
+ *
+ * @category PHP-SDK
+ * @package  Hitmeister\Component\Api\Endpoints\ShippingGroups
+ * @author   Alex Litvinenko <alex.litvinenko@hitmeister.de>
+ * @license  https://opensource.org/licenses/MIT MIT
+ * @link     https://www.hitmeister.de/api/v1/
+ */
+class Find extends AbstractEndpoint
+{
+	use RequestGet;
+
+	/**
+	 * {@inheritdoc}
+	 */
+	public function getParamWhiteList()
+	{
+		return ['limit', 'offset'];
+	}
+
+	/**
+	 * {@inheritdoc}
+	 */
+	public function getURI()
+	{
+		return 'shipping-groups/seller/';
+	}
+}