+++ src/Endpoints/Warehouses/Post.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace Hitmeister\Component\Api\Endpoints\Warehouses;
+
+use Hitmeister\Component\Api\Endpoints\AbstractEndpoint;
+use Hitmeister\Component\Api\Endpoints\Traits\BodyTransfer;
+use Hitmeister\Component\Api\Endpoints\Traits\EmptyParamWhiteList;
+use Hitmeister\Component\Api\Endpoints\Traits\RequestPost;
+use Hitmeister\Component\Api\Transfers\WarehouseAddTransfer;
+
+/**
+ * Class Post
+ *
+ * @category PHP-SDK
+ * @package  Hitmeister\Component\Api\Endpoints\Warehouses
+ * @author   Alex Litvinenko <alex.litvinenko@hitmeister.de>
+ * @license  https://opensource.org/licenses/MIT MIT
+ * @link     https://www.hitmeister.de/api/v1/
+ */
+class Post extends AbstractEndpoint
+{
+	use RequestPost;
+	use EmptyParamWhiteList;
+	use BodyTransfer;
+
+	/**
+	 * @param WarehouseAddTransfer $transfer
+	 */
+	public function setTransfer(WarehouseAddTransfer $transfer)
+	{
+		$this->transfer = $transfer;
+	}
+
+	/**
+	 * {@inheritdoc}
+	 */
+	public function getURI()
+	{
+		return 'warehouses/';
+	}
+}