+++ src/Namespaces/ProductDataNamespace.php
@@ -2,11 +2,11 @@
 
 namespace Hitmeister\Component\Api\Namespaces;
 
-use Hitmeister\Component\Api\Endpoints\ProductData\Get;
-use Hitmeister\Component\Api\Endpoints\ProductData\Put;
-use Hitmeister\Component\Api\Endpoints\ProductData\Update;
 use Hitmeister\Component\Api\Endpoints\ProductData\Delete;
-use Hitmeister\Component\Api\Exceptions\InvalidArgumentException;
+use Hitmeister\Component\Api\Endpoints\ProductData\Get;
+use Hitmeister\Component\Api\Endpoints\ProductData\Upsert;
+use Hitmeister\Component\Api\Endpoints\ProductData\Update;
+use Hitmeister\Component\Api\Exceptions\ResourceNotFoundException;
 use Hitmeister\Component\Api\Namespaces\Traits\PerformWithId;
 use Hitmeister\Component\Api\Transfers\ProductDataTransfer;
 
@@ -24,66 +24,60 @@ class ProductDataNamespace extends AbstractNamespace
 	use PerformWithId;
 
 	/**
-	 * @param string   $ean
+	 * @param string $ean
+	 *
 	 * @return ProductDataTransfer|null
 	 */
 	public function get($ean)
 	{
 		$endpoint = new Get($this->getTransport());
-
 		$result = $this->performWithId($endpoint, $ean);
+		
 		return $result ? ProductDataTransfer::make($result) : null;
 	}
 
 	/**
-	 * @param string $ean
-	 * @param array|ProductDataTransfer $data
-	 * @return bool
+	 * @param string              $ean
+	 * @param ProductDataTransfer $data
+	 *
+	 * @return string
+	 * @throws \Hitmeister\Component\Api\Exceptions\ServerException
 	 */
-	public function put($ean, $data)
+	public function upsert($ean, ProductDataTransfer $data)
 	{
-		if (!$data instanceof ProductDataTransfer) {
-			if (!is_array($data)) {
-				throw new InvalidArgumentException('Data argument should be an array or instance of ProductDataTransfer');
-			}
-			$data = ProductDataTransfer::make($data);
-		}
-
-		$endpoint = new Put($this->getTransport());
+		$endpoint = new Upsert($this->getTransport());
 		$endpoint->setId($ean);
 		$endpoint->setTransfer($data);
 
-		$result = $endpoint->performRequest();
+		$resultRequest = $endpoint->performRequest();
 
-		$dStatus = $result['status'];
-		return $dStatus == 204 || $dStatus == 201;
+		return $resultRequest['status'] == 201;
 	}
 
 	/**
-	 * @param string $ean
-	 * @param array|ProductDataTransfer $data
+	 * @param string              $id
+	 * @param ProductDataTransfer $data
+	 *
 	 * @return bool
 	 */
-	public function update($ean, $data)
+	public function update($id, ProductDataTransfer $data)
 	{
-		if (!$data instanceof ProductDataTransfer) {
-			if (!is_array($data)) {
-				throw new InvalidArgumentException('Data argument should be an array or instance of ProductDataTransfer');
-			}
-			$data = ProductDataTransfer::make($data);
-		}
-
 		$endpoint = new Update($this->getTransport());
-		$endpoint->setId($ean);
+		$endpoint->setId($id);
 		$endpoint->setTransfer($data);
 
-		$result = $endpoint->performRequest();
-		
+		try {
+			$result = $endpoint->performRequest();
+		} catch (ResourceNotFoundException $e) {
+			return false;
+		}
+
 		return $result['status'] == 204;
 	}
 
 	/**
 	 * @param string $ean
+	 * 
 	 * @return bool
 	 */
 	public function delete($ean)