+++ src/Transfers/ShippingGroupTransfer.php
@@ -1,0 +1,40 @@
+<?php
+
+namespace Hitmeister\Component\Api\Transfers;
+
+/**
+ * This class was auto generated. Please, do not modify it!
+ *
+ * @codeCoverageIgnore
+ *
+ * @property string $name
+ *
+ *
+ */
+class ShippingGroupTransfer extends AbstractTransfer
+{
+    /**
+     * @return array
+     */
+    public function getProperties()
+    {
+        static $properties = array (
+  'name' => 
+  array (
+    'embedded' => false,
+    'is_multiple' => false,
+  ),
+);
+        return $properties;
+    }
+
+    /**
+     * @param array $data
+     *
+     * @return ShippingGroupTransfer
+     */
+    public static function make(array $data)
+    {
+        return AbstractTransfer::makeTransfer('Hitmeister\Component\Api\Transfers\ShippingGroupTransfer', $data);
+    }
+}